perf(test): improve test3
