+++ bin/general/previewer.lua
@@ -87,13 +87,13 @@ elseif metaopts.previewer_type == "command_list" then
         return
     end
 
-    local async_spawn = shell_helpers.AsyncSpawn:make(cmd_splits, println) --[[@as AsyncSpawn]]
+    local asp = shell_helpers.Spawn:make(cmd_splits, println) --[[@as Spawn]]
     shell_helpers.log_ensure(
-        async_spawn ~= nil,
+        asp ~= nil,
         "failed to open async command: %s",
         vim.inspect(cmd_splits)
     )
-    async_spawn:run()
+    asp:run()
 elseif metaopts.previewer_type == "list" then
     local f = io.open(resultfile, "r")
     shell_helpers.log_ensure(
