+++ lua/fzfx/config.lua
@@ -11,7 +11,7 @@ local PreviewerTypeEnum = require("fzfx.schema").PreviewerTypeEnum
 local CommandFeedEnum = require("fzfx.schema").CommandFeedEnum
 
 --- @type table<string, FzfOpt>
-local default_fzf_options = {
+local _default_fzf_options = {
     multi = "--multi",
     toggle = "--bind=ctrl-e:toggle",
     toggle_all = "--bind=ctrl-a:toggle-all",
@@ -1875,7 +1875,7 @@ local Defaults = {
             ["ctrl-q"] = require("fzfx.actions").setqflist_find,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             function()
                 return { "--prompt", path.shorten() .. " > " }
             end,
@@ -2073,7 +2073,7 @@ local Defaults = {
                 or require("fzfx.actions").setqflist_grep,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             "--disabled",
             { "--prompt", "Live Grep > " },
             { "--delimiter", ":" },
@@ -2196,7 +2196,7 @@ local Defaults = {
             ["ctrl-q"] = require("fzfx.actions").setqflist_find,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             { "--prompt", "Buffers > " },
             function()
                 local current_bufnr = vim.api.nvim_get_current_buf()
@@ -2322,7 +2322,7 @@ local Defaults = {
             ["ctrl-q"] = require("fzfx.actions").setqflist_find,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             function()
                 return { "--prompt", path.shorten() .. " > " }
             end,
@@ -2542,7 +2542,7 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").git_checkout,
         },
         fzf_opts = {
-            default_fzf_options.no_multi,
+            _default_fzf_options.no_multi,
             { "--prompt", "GBranches > " },
             function()
                 local cmd = require("fzfx.cmd")
@@ -2706,7 +2706,7 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").yank_git_commit,
         },
         fzf_opts = {
-            default_fzf_options.no_multi,
+            _default_fzf_options.no_multi,
             { "--prompt", "GCommits > " },
         },
     },
@@ -2797,7 +2797,7 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").yank_git_commit,
         },
         fzf_opts = {
-            default_fzf_options.no_multi,
+            _default_fzf_options.no_multi,
             { "--prompt", "GBlame > " },
         },
     },
@@ -2977,7 +2977,7 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").feed_vim_command,
         },
         fzf_opts = {
-            default_fzf_options.no_multi,
+            _default_fzf_options.no_multi,
             "--header-lines=1",
             { "--preview-window", "~1" },
             { "--prompt", "Commands > " },
@@ -3214,7 +3214,7 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").feed_vim_key,
         },
         fzf_opts = {
-            default_fzf_options.no_multi,
+            _default_fzf_options.no_multi,
             "--header-lines=1",
             { "--preview-window", "~1" },
             { "--prompt", "Key Maps > " },
@@ -3357,7 +3357,7 @@ local Defaults = {
             ["ctrl-q"] = require("fzfx.actions").setqflist_rg,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             { "--delimiter", ":" },
             { "--preview-window", "+{2}-/2" },
             { "--prompt", "Diagnostics > " },
@@ -3404,8 +3404,8 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").edit_rg,
         },
         fzf_opts = {
-            default_fzf_options.multi,
-            default_fzf_options.lsp_preview_window,
+            _default_fzf_options.multi,
+            _default_fzf_options.lsp_preview_window,
             "--border=none",
             { "--delimiter", ":" },
             { "--prompt", "Definitions > " },
@@ -3464,8 +3464,8 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").edit_rg,
         },
         fzf_opts = {
-            default_fzf_options.multi,
-            default_fzf_options.lsp_preview_window,
+            _default_fzf_options.multi,
+            _default_fzf_options.lsp_preview_window,
             "--border=none",
             { "--delimiter", ":" },
             { "--prompt", "TypeDefinitions > " },
@@ -3524,8 +3524,8 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").edit_rg,
         },
         fzf_opts = {
-            default_fzf_options.multi,
-            default_fzf_options.lsp_preview_window,
+            _default_fzf_options.multi,
+            _default_fzf_options.lsp_preview_window,
             "--border=none",
             { "--delimiter", ":" },
             { "--prompt", "References > " },
@@ -3584,8 +3584,8 @@ local Defaults = {
             ["double-click"] = require("fzfx.actions").edit_rg,
         },
         fzf_opts = {
-            default_fzf_options.multi,
-            default_fzf_options.lsp_preview_window,
+            _default_fzf_options.multi,
+            _default_fzf_options.lsp_preview_window,
             "--border=none",
             { "--delimiter", ":" },
             { "--prompt", "Implementations > " },
@@ -3749,7 +3749,7 @@ local Defaults = {
             ["double-click"] = edit_file_explorer,
         },
         fzf_opts = {
-            default_fzf_options.multi,
+            _default_fzf_options.multi,
             { "--prompt", path.shorten() .. " > " },
             function()
                 local n = 0
@@ -3784,11 +3784,11 @@ local Defaults = {
         "--layout=reverse",
         "--border=rounded",
         "--height=100%",
-        default_fzf_options.toggle,
-        default_fzf_options.toggle_all,
-        default_fzf_options.toggle_preview,
-        default_fzf_options.preview_half_page_down,
-        default_fzf_options.preview_half_page_up,
+        _default_fzf_options.toggle,
+        _default_fzf_options.toggle_all,
+        _default_fzf_options.toggle_preview,
+        _default_fzf_options.preview_half_page_down,
+        _default_fzf_options.preview_half_page_up,
     },
 
     -- fzf colors
@@ -3953,7 +3953,7 @@ local M = {
     setup = setup,
     get_config = get_config,
     get_defaults = get_defaults,
-    default_fzf_options = default_fzf_options,
+    _default_fzf_options = _default_fzf_options,
 }
 
 return M
