+++ lua/fzfx/shell_helpers.lua
@@ -174,7 +174,7 @@ local M = {
     string_rtrim = require("fzfx.utils").string_rtrim,
     FileLineReader = require("fzfx.utils").FileLineReader,
     readfile = require("fzfx.utils").readfile,
-    AsyncSpawn = require("fzfx.utils").AsyncSpawn,
+    Spawn = require("fzfx.spawn").Spawn,
 }
 
 return M
