+++ test/config_spec.lua
@@ -58,16 +58,16 @@ describe("config", function()
             assert_true(string.len(actual --[[@as string]]) > 0)
         end)
     end)
-    describe("[default_fzf_options]", function()
+    describe("[_default_fzf_options]", function()
         it("is constants", function()
-            assert_eq(conf.default_fzf_options.multi, "--multi")
-            assert_eq(conf.default_fzf_options.toggle, "--bind=ctrl-e:toggle")
+            assert_eq(conf._default_fzf_options.multi, "--multi")
+            assert_eq(conf._default_fzf_options.toggle, "--bind=ctrl-e:toggle")
             assert_eq(
-                conf.default_fzf_options.toggle_all,
+                conf._default_fzf_options.toggle_all,
                 "--bind=ctrl-a:toggle-all"
             )
             assert_eq(
-                conf.default_fzf_options.toggle_preview,
+                conf._default_fzf_options.toggle_preview,
                 "--bind=alt-p:toggle-preview"
             )
         end)
