+++ test/utils_spec.lua
@@ -404,198 +404,6 @@ describe("utils", function()
             assert_eq(utils.string_rtrim(buffer --[[@as string]]), content)
         end)
     end)
-    describe("[AsyncSpawn]", function()
-        it("open", function()
-            local async_spawn = utils.AsyncSpawn:make(
-                { "cat", "README.md" },
-                function() end
-            ) --[[@as AsyncSpawn]]
-            assert_eq(type(async_spawn), "table")
-            assert_eq(type(async_spawn.cmds), "table")
-            assert_eq(#async_spawn.cmds, 2)
-            assert_eq(async_spawn.cmds[1], "cat")
-            assert_eq(async_spawn.cmds[2], "README.md")
-            assert_eq(type(async_spawn.out_pipe), "userdata")
-            assert_eq(type(async_spawn.err_pipe), "userdata")
-        end)
-        it("consume line", function()
-            local content = utils.readfile("README.md") --[[@as string]]
-            local lines = utils.readlines("README.md") --[[@as table]]
-
-            local i = 1
-            local function process_line(line)
-                -- print(string.format("[%d]%s", i, line))
-                assert_eq(type(line), "string")
-                assert_eq(line, lines[i])
-                i = i + 1
-            end
-            local async_spawn =
-                utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-            local pos = async_spawn:_consume_line(content, process_line)
-            if pos <= #content then
-                local line = content:sub(pos, #content)
-                process_line(line)
-            end
-        end)
-        it("stdout on newline", function()
-            local content = utils.readfile("README.md") --[[@as string]]
-            local lines = utils.readlines("README.md") --[[@as table]]
-
-            local i = 1
-            local function process_line(line)
-                -- print(string.format("[%d]%s\n", i, line))
-                assert_eq(type(line), "string")
-                assert_eq(line, lines[i])
-                i = i + 1
-            end
-            local async_spawn =
-                utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-            local content_splits =
-                utils.string_split(content, "\n", { trimempty = false })
-            for j, splits in ipairs(content_splits) do
-                async_spawn:_on_stdout(nil, splits)
-                if j < #content_splits then
-                    async_spawn:_on_stdout(nil, "\n")
-                end
-            end
-            async_spawn:_on_stdout(nil, nil)
-            assert_true(async_spawn.out_pipe:is_closing())
-        end)
-        it("stdout on whitespace", function()
-            local content = utils.readfile("README.md") --[[@as string]]
-            local lines = utils.readlines("README.md") --[[@as table]]
-
-            local i = 1
-            local function process_line(line)
-                -- print(string.format("[%d]%s\n", i, line))
-                assert_eq(type(line), "string")
-                assert_eq(line, lines[i])
-                i = i + 1
-            end
-            local async_spawn =
-                utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-            local content_splits =
-                utils.string_split(content, " ", { trimempty = false })
-            for j, splits in ipairs(content_splits) do
-                async_spawn:_on_stdout(nil, splits)
-                if j < #content_splits then
-                    async_spawn:_on_stdout(nil, " ")
-                end
-            end
-            async_spawn:_on_stdout(nil, nil)
-            assert_true(async_spawn.out_pipe:is_closing())
-        end)
-        for delimiter_i = 0, 25 do
-            -- lower case: a
-            local lower_char = string.char(97 + delimiter_i)
-            it(string.format("stdout on %s", lower_char), function()
-                local content = utils.readfile("README.md") --[[@as string]]
-                local lines = utils.readlines("README.md") --[[@as table]]
-
-                local i = 1
-                local function process_line(line)
-                    -- print(string.format("[%d]%s\n", i, line))
-                    assert_eq(type(line), "string")
-                    assert_eq(line, lines[i])
-                    i = i + 1
-                end
-                local async_spawn =
-                    utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-                local content_splits = utils.string_split(
-                    content,
-                    lower_char,
-                    { trimempty = false }
-                )
-                for j, splits in ipairs(content_splits) do
-                    async_spawn:_on_stdout(nil, splits)
-                    if j < #content_splits then
-                        async_spawn:_on_stdout(nil, lower_char)
-                    end
-                end
-                async_spawn:_on_stdout(nil, nil)
-                assert_true(async_spawn.out_pipe:is_closing())
-            end)
-            -- upper case: A
-            local upper_char = string.char(65 + delimiter_i)
-            it(string.format("stdout on %s", upper_char), function()
-                local content = utils.readfile("README.md") --[[@as string]]
-                local lines = utils.readlines("README.md") --[[@as table]]
-
-                local i = 1
-                local function process_line(line)
-                    -- print(string.format("[%d]%s\n", i, line))
-                    assert_eq(type(line), "string")
-                    assert_eq(line, lines[i])
-                    i = i + 1
-                end
-                local async_spawn =
-                    utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-                local content_splits = utils.string_split(
-                    content,
-                    upper_char,
-                    { trimempty = false }
-                )
-                for j, splits in ipairs(content_splits) do
-                    async_spawn:_on_stdout(nil, splits)
-                    if j < #content_splits then
-                        async_spawn:_on_stdout(nil, upper_char)
-                    end
-                end
-                async_spawn:_on_stdout(nil, nil)
-                assert_true(async_spawn.out_pipe:is_closing())
-            end)
-        end
-        it("stderr", function()
-            local async_spawn = utils.AsyncSpawn:make(
-                { "cat", "README.md" },
-                function() end
-            ) --[[@as AsyncSpawn]]
-            async_spawn:_on_stderr(nil, nil)
-            assert_true(async_spawn.err_pipe:is_closing())
-        end)
-        it("iterate on README.md", function()
-            local lines = utils.readlines("README.md") --[[@as table]]
-
-            local i = 1
-            local function process_line(line)
-                print(string.format("[%d]%s\n", i, line))
-                assert_eq(type(line), "string")
-                assert_eq(lines[i], line)
-                i = i + 1
-            end
-
-            local async_spawn =
-                utils.AsyncSpawn:make({ "cat", "README.md" }, process_line) --[[@as AsyncSpawn]]
-            async_spawn:run()
-        end)
-        it("iterate on lua/fzfx/config.lua", function()
-            local lines = utils.readlines("lua/fzfx/config.lua") --[[@as table]]
-
-            local i = 1
-            local function process_line(line)
-                print(string.format("[%d]%s\n", i, line))
-                assert_eq(type(line), "string")
-                assert_eq(lines[i], line)
-                i = i + 1
-            end
-
-            local async_spawn = utils.AsyncSpawn:make(
-                { "cat", "lua/fzfx/config.lua" },
-                process_line
-            ) --[[@as AsyncSpawn]]
-            async_spawn:run()
-        end)
-        it("close handle", function()
-            local async_spawn = utils.AsyncSpawn:make(
-                { "cat", "lua/fzfx/config.lua" },
-                function() end
-            ) --[[@as AsyncSpawn]]
-            async_spawn:run()
-            assert_true(async_spawn.process_handle ~= nil)
-            async_spawn:_close_handle(async_spawn.process_handle)
-            assert_true(async_spawn.process_handle:is_closing())
-        end)
-    end)
     describe("[list_index]", function()
         it("positive", function()
             for i = 1, 10 do
